+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ### Version 1.2.5
 methods added:
+	+ (void)updateShouldHideOnTap:(BOOL)hide;
+	+ (void)updateShouldShowCloseIcon:(BOOL)show;
+properties added:
+	@property (nonatomic, assign) EHBoolean shouldHideOnTap;
+	@property (nonatomic, assign) EHBoolean shouldShowCloseIcon;
+
+
+### Version 1.2.5
+methods added:
 	-hide
 	+hideAll:animated